FOR-57 - Criado componente para realizar upload de templates.:

+++ app/dashboard/survey-templates/components/survey-template/survey-template-component.js
@@ -23,7 +23,9 @@
         self.isSelected = false;
 
         self.$onDestroy = function() {
-            SelectedSurveyTemplatesManagementService.removeSurveyTemplate(self.surveyTemplate);
+            if (self.isSelected) {
+                SelectedSurveyTemplatesManagementService.removeSurveyTemplate(self.surveyTemplate);
+            }
         };
 
         $element.on('click', function() {
